Move util functions to server util file, user server config to configure download limit
